fmcomms2: Set dac_unpack channels to 4
There are only 4 DAC channels in the fmcomms2 design, so set the number of
channels of the dac_unpack core to 4. This slightly reduces resource usage
as well as reducing the DMA alignment requirement from 128bit to 64bit.  The
later value is what existing applications expect the alignement requirement
to be.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>