common: Place HDMI interface registers into the IOB
The paths from the HDMI interface registers to the IO pads are
unconstrained. This means the P&R can in theory put the register anywhere
which could lead to stability issues on the interface, depending on what
else is in the fabric. To get predictable delays for the register to IO pad
path place the register into the IOB section.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>