axi_dmac: Add support for chained DMA descriptors
This commit introduces a different interface to submit transfers, using
DMA descriptors.

The structure of the DMA descriptor must be as follows:

struct dma_desc {
    u32 flags,
    u32 id,
    u64 dest_addr,
    u64 src_addr,
    u64 next_sg_addr,
    u32 y_len,
    u32 x_len,
    u32 src_stride,
    u32 dst_stride,
};

The 'flags' field currently offers two control bits:
- bit 0: if set, the transfer will complete after this last descriptor
  is processed, and the DMA core will go back to idle state.
  If clear, the next DMA descriptor (which address is contained
  in 'next_sg_addr') will be loaded.
- bit 1: if set, an end-of-transfer interrupt will be raised after the
  transfer corresponding to this descriptor completed.

The 'id' field should correspond to an identifier of the descriptor.
When a descriptor has been loaded, register 0x440 is updated with the
value of this field.

The 'dest_addr' and 'src_addr' should contain the destination and source
addresses to use for the transfer, respectively.

The 'x_len' field should contain the amount of bytes to transfer,
minus one.

The 'y_len', 'src_stride' and 'dst_stride' fields are only useful for
2D transfers, and should be set to zero if 2D transfers are not
required.

To start a transfer, the address of the first DMA descriptor should be
written to register 0x43c.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>