axi_ad9144: Completely disable unused channels in DUAL mode
When the axi_ad9144 core is configured for DUAL mode two of the four
channels are unused. But there is still some residual logic left for those
unused channels that can't be removed by the optimizer.

Completely disable the unused channels by reducing the channel and lane
count. This slightly reduces utilization.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>